requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
Flask==2.0.2
Flask-RESTful==0.3.9
fonttools==4.28.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
imageio==2.13.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scipy==1.7.3
setuptools-scm==6.3.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
